refactor: 외부 api 호출 Transactional 밖에서
